move common enums to common document ai module
